Move Main to its own file
bus.rb was way too big
